requirements.txt
asgiref==3.3.1
Django==3.1.4
django-resized==0.3.11
libmagic==1.0
Pillow==8.0.1
python-magic==0.4.18
pytz==2020.4
sqlparse==0.4.1
whitenoise==5.2.0
